+++ PSLab_Original/PSLAB_ADC.h
@@ -9,7 +9,7 @@
 #define	PSLAB_ADC_H
 
 extern BYTE CHOSA, CH123SA;
-extern unsigned int ADC_DELAY;
+extern uint16 ADC_DELAY;
 extern BYTE conversion_done;
 extern BYTE TRIGGERED;
 extern BYTE TRIGGER_READY;
@@ -17,9 +17,9 @@ extern BYTE TRIGGER_CHANNEL;
 extern int *buff0, *buff1, *endbuff, *buff2, *buff3;
 extern BYTE ADC_CHANNELS; // CH1 only
 extern int samples;
-extern unsigned int samples_to_fetch;
-extern unsigned int adval;
-extern unsigned int TRIGGER_TIMEOUT, TRIGGER_WAITING, TRIGGER_LEVEL, TRIGGER_PRESCALER;
+extern uint16 samples_to_fetch;
+extern uint16 adval;
+extern uint16 TRIGGER_TIMEOUT, TRIGGER_WAITING, TRIGGER_LEVEL, TRIGGER_PRESCALER;
 
 extern void EnableComparator();
 extern void DisableComparator();
@@ -30,8 +30,8 @@ extern void initADC12(void);
 extern void initADCCTMU(void);
 extern void initADC12_averaging16();
 extern void setADCMode(BYTE, BYTE, BYTE);
-extern unsigned int get_voltage_summed(BYTE channel);
-extern unsigned int get_voltage(BYTE channel);
+extern uint16 get_voltage_summed(BYTE channel);
+extern uint16 get_voltage(BYTE channel);
 extern void setupADC10();
 extern void configureADC();
 extern void enableADCDMA();
