+++ PSLab_Original/Wave_Generator.h
@@ -2,7 +2,7 @@
  * File:   Wave_Generator.h
  * Author: padmal
  *
- * Created on June 11, 2017, 3:49 AM
+ * Created on June 10, 2017, 3:49 AM
  */
 
 #ifndef WAVE_GENERATOR_H
@@ -13,12 +13,12 @@ extern int __attribute__((section("sine_table2"))) sineTable2[];
 extern int __attribute__((section("sine_table1_short"))) sineTable1_short[];
 extern int __attribute__((section("sine_table2_short"))) sineTable2_short[];
 
-extern void sqr1(unsigned int, unsigned int, BYTE);
-extern void sqr2(unsigned int, unsigned int, BYTE);
+extern void sqr1(uint16, uint16, BYTE);
+extern void sqr2(uint16, uint16, BYTE);
 extern void sqr4(uint16 w,uint16 R0,uint16 R1,uint16 RS1,uint16 R2,uint16 RS2,uint16 R3,uint16 RS3,BYTE scaling);
-extern void sineWave1(unsigned int wavelength,BYTE highres);
-extern void sineWave2(unsigned int wavelength,BYTE highres);
-extern void setSineWaves(unsigned int wavelength1,unsigned int wavelength2,unsigned int pos,unsigned int tmr_delay,BYTE highres);
+extern void sineWave1(uint16 wavelength,BYTE highres);
+extern void sineWave2(uint16 wavelength,BYTE highres);
+extern void setSineWaves(uint16 wavelength1,uint16 wavelength2,uint16 pos,uint16 tmr_delay,BYTE highres);
 
 #endif	/* WAVE_GENERATOR_H */
